macho: add some objcopy tests
Currently only x86-64 and AArch64 work.

i386 and ARM are missing support for scattered relocations and
paired relocations.

Also add support for an environment called OBJECT_TESTFILES_UPDATE
which causes the tests to update the output files.
