+++ crates/examples/src/lib.rs
@@ -1,6 +1,9 @@
 // Style.
 #![allow(clippy::single_match)]
 
+#[cfg(all(feature = "read", feature = "write"))]
+pub mod objcopy;
+
 #[cfg(feature = "read")]
 pub mod objdump;
