+++ crates/examples/testfiles/macho/base-aarch64.o.objcopy
@@ -1,0 +1,164 @@
+Format: Mach-O 64-bit
+MachHeader {
+    Magic: 0xCFFAEDFE
+    CpuType: CPU_TYPE_ARM64 (0x100000C)
+    CpuSubtype: 0x0
+        CPU_SUBTYPE_ARM64_ALL (0x0)
+    FileType: MH_OBJECT (0x1)
+    NumberOfCmds: 2
+    SizeOfCmds: 0x150
+    Flags: MH_SUBSECTIONS_VIA_SYMBOLS (0x2000)
+}
+SegmentCommand {
+    Cmd: LC_SEGMENT_64 (0x19)
+    CmdSize: 0x138
+    SegmentName: ""
+    VmAddress: 0x0
+    VmSize: 0x68
+    FileOffset: 0x170
+    FileSize: 0x68
+    MaxProt: 0x7
+        VM_PROT_READ (0x1)
+        VM_PROT_WRITE (0x2)
+        VM_PROT_EXECUTE (0x4)
+    InitProt: 0x7
+        VM_PROT_READ (0x1)
+        VM_PROT_WRITE (0x2)
+        VM_PROT_EXECUTE (0x4)
+    NumberOfSections: 3
+    Flags: 0x0
+    Section {
+        Index: 0
+        SectionName: "__text"
+        SegmentName: "__TEXT"
+        Address: 0x0
+        Size: 0x34
+        Offset: 0x170
+        Align: 0x2
+        RelocationOffset: 0x260
+        NumberOfRelocations: 0x3
+        Flags: 0x80000400
+            S_REGULAR (0x0)
+            S_ATTR_PURE_INSTRUCTIONS (0x80000000)
+            S_ATTR_SOME_INSTRUCTIONS (0x400)
+        RelocationInfo {
+            Address: 0x20
+            Extern: yes
+            Symbol: "_printf" (0x5)
+            PcRel: yes
+            Length: 2
+            Type: ARM64_RELOC_BRANCH26 (0x2)
+        }
+        RelocationInfo {
+            Address: 0x1C
+            Extern: yes
+            Symbol: "l_.str" (0x1)
+            PcRel: no
+            Length: 2
+            Type: ARM64_RELOC_PAGEOFF12 (0x4)
+        }
+        RelocationInfo {
+            Address: 0x18
+            Extern: yes
+            Symbol: "l_.str" (0x1)
+            PcRel: yes
+            Length: 2
+            Type: ARM64_RELOC_PAGE21 (0x3)
+        }
+    }
+    Section {
+        Index: 1
+        SectionName: "__cstring"
+        SegmentName: "__TEXT"
+        Address: 0x34
+        Size: 0xD
+        Offset: 0x1A4
+        Align: 0x0
+        RelocationOffset: 0x0
+        NumberOfRelocations: 0x0
+        Flags: S_CSTRING_LITERALS (0x2)
+    }
+    Section {
+        Index: 2
+        SectionName: "__compact_unwind"
+        SegmentName: "__LD"
+        Address: 0x48
+        Size: 0x20
+        Offset: 0x1B8
+        Align: 0x3
+        RelocationOffset: 0x278
+        NumberOfRelocations: 0x1
+        Flags: 0x2000000
+            S_REGULAR (0x0)
+            S_ATTR_DEBUG (0x2000000)
+        RelocationInfo {
+            Address: 0x0
+            Extern: no
+            Section: "__TEXT,__text" (0x1)
+            PcRel: no
+            Length: 3
+            Type: ARM64_RELOC_UNSIGNED (0x0)
+        }
+    }
+}
+SymtabCommand {
+    Cmd: LC_SYMTAB (0x2)
+    CmdSize: 0x18
+    SymbolOffset: 0x1D8
+    NumberOfSymbols: 0x6
+    StringOffset: 0x238
+    StringSize: 0x28
+    Nlist {
+        Index: 0
+        String: "ltmp0" (0x22)
+        Type: N_SECT (0xE)
+        Section: "__TEXT,__text" (0x1)
+        Desc: 0x0
+        Value: 0x0
+    }
+    Nlist {
+        Index: 1
+        String: "l_.str" (0x1)
+        Type: N_SECT (0xE)
+        Section: "__TEXT,__cstring" (0x2)
+        Desc: 0x0
+        Value: 0x34
+    }
+    Nlist {
+        Index: 2
+        String: "ltmp1" (0x1C)
+        Type: N_SECT (0xE)
+        Section: "__TEXT,__cstring" (0x2)
+        Desc: 0x0
+        Value: 0x34
+    }
+    Nlist {
+        Index: 3
+        String: "ltmp2" (0x16)
+        Type: N_SECT (0xE)
+        Section: "__LD,__compact_unwind" (0x3)
+        Desc: 0x0
+        Value: 0x48
+    }
+    Nlist {
+        Index: 4
+        String: "_main" (0x8)
+        Type: 0xF
+            N_SECT (0xE)
+            N_EXT (0x1)
+        Section: "__TEXT,__text" (0x1)
+        Desc: 0x0
+        Value: 0x0
+    }
+    Nlist {
+        Index: 5
+        String: "_printf" (0xE)
+        Type: 0x1
+            N_UNDF (0x0)
+            N_EXT (0x1)
+        Section: "" (0x0)
+        Desc: 0x0
+            REFERENCE_FLAG_UNDEFINED_NON_LAZY (0x0)
+        Value: 0x0
+    }
+}
